backend-functions/requirements.txt
azure-functions
azure-storage-blob
Pillow
PyPDF2
moviepy
numpy